+++ src/routes/JobSummary/components/index.js
@@ -1,0 +1,2 @@
+export * from './RecordsTable';
+export * from './SummaryTable';